+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Trim spaces from both sides of the attribute value.
 - Hash impl for enums.
 - All *Units attribute values are parsed now.
+- The `InvalidLength` error.
 
 ### Changed
 - Rename `AdvanceError` into `InvalidAdvance`.